Remove unreachable post-redirect user creation code
signInWithRedirect returns Promise<void>, so I'm pretty sure this code is unreachable and wouldn't function if it were, we don't notice because the page redirects before the removed code can run.

The reason it works is because onAuthChange fires when the page reloads from the redirect.

We could explicitly handle the redirect, but I'm not sure if there's any advantage vs. waiting for onAuthChange to fire (it always fires on first page load, with a null userCredential if not logged in)